requirements.txt
pandas~=1.3.0
requests~=2.26.0
bs4~=0.0.1
beautifulsoup4~=4.9.3
openpyxl~=3.0.9
lxml~=4.6.3

python-dateutil>=2.8.2
xlrd>=1.2.0